% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{Function that builds a layer based on the specification of the GeomHurricane class.}
\usage{
geom_hurricane(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{data}{A dataframe containing hurricane data. Should include latitude and longitude
data for the center of the hurricane as well as radii values indicating how far from the center
of the hurricane certain windspeeds extend in the NE, NW, SW and SE directions.}
}
\description{
Function that builds a layer based on the specification of the GeomHurricane class.
}
\examples{
  \dontrun{ggplot(data = katrina) +
               geom_hurricane(aes(x = longitude, y = latitude,
                               r_ne = ne, r_se = se, r_nw = nw, r_sw = sw,
                               fill = wind_speed, color = wind_speed)) +
               scale_color_manual(name = "Wind speed (kts)",
                                  values = c("red", "orange", "yellow")) +
               scale_fill_manual(name = "Wind speed (kts)",
                                 values = c("red", "orange", "yellow"))}

}
