% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{Defines the GeomHurricane class via the ggproto function in ggplot2.}
\format{
An object of class \code{GeomHurricane} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomHurricane
}
\description{
Defines the GeomHurricane class via the ggproto function in ggplot2.
}
\keyword{datasets}
