% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_script.R
\name{get_hurricane_data}
\alias{get_hurricane_data}
\title{Reads in and cleans raw hurricane data for use with our new ggplot geom}
\usage{
get_hurricane_data()
}
\description{
Reads in and cleans raw hurricane data for use with our new ggplot geom
}
\examples{
\dontrun{katrina_data <- get_hurricane_data() \%>\%
                         filter(startsWith(storm_id, "KATRINA")}

}
